feat: use nodesource for Node.js installation
Replaced snap installation with nodesource setup for Node.js.
For WSL (Windows Subsystem for Linux) and Docker environments, it is
recommended to use the NodeSource repository when installing Node.js
because it is difficult to run `snapd` with `systemd`. This method
allows you to easily install and manage the latest version of Node.js.

https://github.com/nodesource/distributions?tab=readme-ov-file#installation-instructions-deb